+++ node/src/command.rs
@@ -432,4 +432,8 @@ impl CliConfiguration<Self> for RelayChainCli {
 	) -> Result<Option<sc_telemetry::TelemetryEndpoints>> {
 		self.base.base.telemetry_endpoints(chain_spec)
 	}
+
+	fn node_name(&self) -> Result<String> {
+		self.base.base.node_name()
+	}
 }
